环境搭建成功。

+++ src/main/java/com/jblog/repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.jblog.repository;
 import com.jblog.domain.article.Article;
 
 public interface ArticleRepository {
-    Article getArticle(int id);
+    Article getArticle(long id);
     int count();
 }
